moved message implementation
