+++ pyodsp/dec/bd/alg_leaf_pyomo.py
@@ -6,8 +6,7 @@ import pandas as pd
 from pyomo.environ import Suffix
 from pyomo.core.base.constraint import ScalarConstraint
 
-from pyodsp.dec.run._message import BdInitMessage, BdFinalMessage
-
+from .message import BdInitMessage, BdFinalMessage
 from .alg_leaf import BdAlgLeaf
 from ..utils import CouplingData, get_nonzero_coefficients_from_model
 from pyodsp.alg.cuts import Cut, OptimalityCut, FeasibilityCut
